+++ crates/swc_ecma_transforms_typescript/tests/fixture/issue-8007/2/input.ts
@@ -1,0 +1,9 @@
+import { A } from 'mod';
+
+const foo = <
+  T extends ((A: any) => void),
+  >(
+    arg: T,
+) => arg;
+
+console.log(A);