+++ crates/swc_ecma_transforms_typescript/tests/fixture/issue-8007/2/output.js
@@ -1,0 +1,3 @@
+import { A } from 'mod';
+const foo = (arg)=>arg;
+console.log(A);